decouple add cost pt1
